Module Nottui.Gravity

type direction = [

| `Negative

| `Neutral

| `Positive

]

A gravity is a pair of directions along the horizontal and vertical axis.

Horizontal axis goes from left to right and vertical axis from top to bottom.

`Negative direction means left / top bounds, `Neutral means center and `Positive means right / bottom.

val pp_direction : Stdlib.Format.formatter -> direction -> unit

Printing directions

type t

The gravity type is a pair of an horizontal and a vertical gravity

val pp : Stdlib.Format.formatter -> t -> unit

Printing gravities

val make : h:direction -> v:direction -> t

Make a gravity value from an horizontal and a vertical directions.

val default : t

Default (negative, aligning to the top-left) gravity.

val h : t -> direction

Get the horizontal direction

val v : t -> direction

Get the vertical direction